+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>joystick_operator</name>
+  <name>umrt_arm_joystick_operator</name>
   <version>1.0.0</version>
   <description>Joystick teleop for the University of Manitoba Robotics Team's Robotic Arm</description>
   <maintainer email="noah.reeder@ieee.org">Noah Reeder</maintainer>
